+++ cpp/runtimes/libjoynr-runtime/websocket/JoynrRuntime.cpp
@@ -24,14 +24,15 @@
 namespace joynr
 {
 
-JoynrRuntime* JoynrRuntime::createRuntime(const std::string& pathToLibjoynrSettings,
-                                          const std::string& pathToMessagingSettings)
+std::unique_ptr<JoynrRuntime> JoynrRuntime::createRuntime(
+        const std::string& pathToLibjoynrSettings,
+        const std::string& pathToMessagingSettings)
 {
 
     return createRuntime(createSettings(pathToLibjoynrSettings, pathToMessagingSettings));
 }
 
-JoynrRuntime* JoynrRuntime::createRuntime(std::unique_ptr<Settings> settings)
+std::unique_ptr<JoynrRuntime> JoynrRuntime::createRuntime(std::unique_ptr<Settings> settings)
 {
     Future<std::unique_ptr<JoynrRuntime>> runtimeFuture;
 
@@ -50,7 +51,7 @@ JoynrRuntime* JoynrRuntime::createRuntime(std::unique_ptr<Settings> settings)
     std::unique_ptr<JoynrRuntime> runtime;
     runtimeFuture.get(runtime);
 
-    return runtime.release();
+    return runtime;
 }
 
 void JoynrRuntime::createRuntimeAsync(